Cleanup temp directory created by GitRepository
Change the interface of GitRepository so that it accepts a directory in
its initializer. The class will force Git to execute in this directory.

Add a new `.temporary` class methods that instantiates a GitRepository
instance with a temp directory and yields the instance. Once the block
terminates, the temp directory is removed.