Update AddRepository component
Change the form action in AddRepository.tsx to make a request to the
repository search API endpoint.

If the repository exists in our database, we navigate to
`/repository/:repository_id/change-volume`.

If the repository exists on the remote, but not in our database, we show
an alert to the user that the repository exists, but not yet analyzed.

If the repository does not exists anywhere, we just mention that to the
user.