Categorize type of change in RepositorySyncService
Update the RepositorySyncService to set the `category` column on each
`source_file_changes` rows it creates to the type of change returned by
`git log`.

Change that creates a file are tagged with the `create` category.
Change that modifies a file are tagged with the `modify` category.
Change that deletes a file are tagged with the `delete` category.