Fix bug in RepositorySyncService
Fix a bug where the ledger commits were not extracted during sync.

The sync service runs in two steps:
1. it extracts all commits in the history
2. it extracts "ledger" commits, commits as they appear on the "main"
   branch of the repository.

In https://github.com/visevol/Cscope/pull/73, we introduced the concept
of "last known commit" in the RepositorySyncService. The idea is to
start the extraction process from the latest commit in the database.

The bug is that the "last know commit" is fetched in both steps.

On the first step, the last known commit is accurate. For new
repositories, we don't have any commits in our database, so the
`git log` command extracts all commits, and we store this in the
database.

On the second step, the last known commit exists because it was written
on the first step. Here, `git log` returns nothing because we're
essentially calling `git log HEAD..HEAD`.